added role based validation and updated the account controller to use it
