+++ 3DMultiplayerGame/views/Account/register.vash
@@ -8,7 +8,7 @@
         
         <div class="row">
             <div class="col-sm-5 bg-info text-light" style="margin: 100px auto 100px;border-radius: 5px;">
-                <form class="form" method="POST" action="/api/Account/register" style="padding: 3vh;">
+                <form class="form" method="POST" action="/Account/register.html" style="padding: 3vh;">
                     <h3>Register</h3>
                     <div class="form-group">
                         @html.LabelFor(model.username)
@@ -17,12 +17,12 @@
                     </div>
                     <div class="form-group">
                         @html.LabelFor(model.password)
-                        @html.PasswordBoxFor(model.password, { class: "form-control", autocomplete: "off"})
+                        @html.PasswordBoxFor(model.password, { class: "form-control", autocomplete: "new-password"})
                         @html.ValidationMessageFor(model.password)
                     </div>
                     <div class="form-group">
                         @html.LabelFor(model.passwordConfirmation)
-                        @html.PasswordBoxFor(model.passwordConfirmation, { class: "form-control", autocomplete: "off"})
+                        @html.PasswordBoxFor(model.passwordConfirmation, { class: "form-control", autocomplete: "new-password"})
                         @html.ValidationMessageFor(model.passwordConfirmation)
                     </div>
                     <div>
