+++ 3DMultiplayerGame/views/Shared/_layout.vash
@@ -50,7 +50,7 @@
                             <a class="btn btn-outline-info" href="/Account/dashboard.html">Dashboard</a>
                         </li>
                         <li class="nav-item">
-                            <a class="btn btn-outline-info" href="/api/Account/logout">Logout</a>
+                            <a class="btn btn-outline-info" href="/Account/logout.html">Logout</a>
                         </li>
                         }
